Manejo de errores en inicio de sesion

+++ src/main/java/dao/AlumnoDAO.java
@@ -21,7 +21,7 @@ public interface AlumnoDAO {
     public ArrayList<ProfesorDTO> listarProfesores(int idAlumno);
     public ArrayList<AlumnoDTO> listarAlumnos(int idAlumno);
     public ArrayList<AsignaturaDTO> listarAsignaturas(int idAlumno);
-    public ArrayList<Integer> listarNotasAsignatura(int idAsign, int idAlum);
+    public ArrayList<Double> listarNotasAsignatura(int idAsign, int idAlum);
     
     //Para el admin
     public List listar();
